requirements.txt
sqlmodel
fastapi
uvicorn
psycopg2-binary
python-dateutil
python-dotenv
openai
twilio
